refactor(magic): split types.py into thematic submodules
Prepare for Scope 6 new dataclasses (RitualOutcome, etc.) by converting
types.py into a package with thematic submodules per §4.4 / §7.
