+++ src/world/magic/types/ritual.py
@@ -1,0 +1,12 @@
+from __future__ import annotations
+
+from django.db import models
+
+
+class AnimaRitualCategory(models.TextChoices):
+    """Categories of anima recovery rituals."""
+
+    SOLITARY = "solitary", "Solitary"
+    COLLABORATIVE = "collaborative", "Collaborative"
+    ENVIRONMENTAL = "environmental", "Environmental"
+    CEREMONIAL = "ceremonial", "Ceremonial"
